Add representable framework
Allow custom representations of objects passed to the Liquid context.
The representation will automatically convert to string keys so that the
object becomes accessible to the template.

Customization is possible by implementing the `Liquex.Representable`
protocol.  This includes choosing which fields to represent and allowing
lazy invocation of the representation.